Fix TEP Linting
This change fixes the date on TEP-0089 to pass the PipelineRun on
open TEPs.

The `pull-community-teps-lint` PipelineRun did not run when we
merged TEP-0089 in https://github.com/tektoncd/community/pull/529.
It got merged with a linting issue - wrong date - that's causing
the same PipelineRun to fail in open TEPs.