TEP-0025: API change - use string alias instead of bool
Today, hermetic execution is supported through annotations as an
experiment. If successful, we plan to support hermetic execution
through a field in the API. The proposed API uses a bool, however
the Kubernetes API conventions [discourage using booleans][bools]
because they limit future expansions. This change updates the field
to use a string alias instead, as recommended in the conventions.

[bools]: https://github.com/kubernetes/community/blob/fbc5ca53d6bfd8388b335f7d0198b67a14b99d91/contributors/devel/sig-architecture/api-conventions.md#primitive-types