[TEP-0086] Update requirements and alternatives after recent discussion 📓
In [a recent API working group meeting](https://docs.google.com/document/d/17PodAxG8hV351fBhSu7Y_OIPhGTVgj6OJ2lPphYYRpU/edit#heading=h.ux4qhojeot6k)
we talked through the proposal that was currently looking the most
appealing (sidecar + HTTP service) and @skaegi brought up some potential
requirements that might change this, and also volunteered to find
someone to work on a POC for a logs based solution before we disregard
that entirely.

Personally I agree that it would be best if we could keep the TaskRun to
be the source of truth for the results themselves, but that would mean
agreeing to an upper limit on result size (at least).

I've tried to reflect the above and generally organize the TEP a bit,
and add some missing stuff:
* Moved some points from Goals to requirements (really one goal overall,
  rest of the items are constraints on how we accomplish that goal)
* Add requirements we've been discussing, trying to highlight the ones
  that are newer and other collaborators might disagree with
* Expanding the risks section a bit
* Added some high level performance details
* Suggested potential next steps to unblock this
* Added more detials to auth/encryption/defined interface notes from
  our previous one off meeting
  [notes](https://hackmd.io/YU_g27vRS2S5DwfBXDGpYA?view)
* Filled in the design evaluation
* Added some drawbacks
* Separated "SideCar Result References" into two alternatives (which are
  not mutually exclusive)
* Added using PVCs as an alternative
* Added some cons to log based solution