Experiment with removing team maintainers
GitHub let us define a list of team members and team maintainers.
Since we maintain teams through peribolos, we should not really
have a list of team maintainers.

Each repo has a <repo>.maintainers and a <repo>.<collaborators>
team, each with its own list of maintainers and members, which
make things notably more confusing.

Last and not least, we have a pending task of sync'ing teams to
OWNER files in repos, however today it's hard to be sure that
the list of people members of <repo>.maintainers teams are those
who need to be in OWNERS since each individual can only be a
maintainer or a member.

Start experimenting doing this for the catalog team.

This also means that the Tekton governance team members do not
need to be members of all teams and maintainers of all repos.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>