Add a leave policy for governing board members 🏖️
Since I am soon going to be on leave, I brought this up in a recent
governing board / community meeting, and this commit adds the policy
that we discussed adding for governing board members going on leave in
general. We wanted to make sure that this option was available
regardless of the reason for the leave.

The maximum period of time for the leave is however conveniently the
exact length of time I plan to be gone XD , so if folks have different
opinions on what the threshold should be, we can definitely change it