Drop the proposal and update requirements
I noticed this TEP is not yet accepted and merged. This TEP is stuck in the
review process. Dropping the proposal and design details based on the
recent discussion in the API WG. This TEP instead now contains a list of
alternatives. If we can merge this without proposal, it opens up an opportunity
for the community to create a PoC and evaluate which solution fits best for
the set of the requirements.

Moving the proposal and design details under the first alternative, making it
more like a TEP within TEP.

One major change here is the requirement section in which task results as part
of the taskRun status is listed as a requirement with the flexible design for
the use cases which would like to avoid it.

Updated the toc.