TEP-0101: Decouple catalog organization and reference
As of today, users of the tekton task are directly refering to the
GitHub URL to get Task from the catalog. This creates a hard coupling
between the way we organize tasks inside Tekton and the way user
consume it. This is forcing us into one possible catalog organization
and prevent us to make drastic changes to this organization without
breaking users.

This enhancement aim to decouple the catalog organization from users
consumption of those resources. This enhancement proposal is mainly
touching the catalog and the hub components. This would also push the
hub to be the main entrypoint for users to search, list and get tekton
resources.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>