[TEP-0075, 0076] Required by default 📜
After recent discussions in API working group, landing on implying the
required field by default and not supporting users explicitly specifying
required fields - seems to be the best balance of the functionality we
want and setting us up to be able to change the behavior later if
needed.

Added another alternative for how to handle required vs optional:
if we provided a default value when optional fields are missing

Added future work where we could support "no schema" i.e. objects that
could have any keys and declare none.