[TEP-0074] Mark as implementable
This commit updates TEP-0074 to reflect that implementing array and object support, plus
a solution for data locality, are not blockers for removing PipelineResources. It proposes
that PipelineResources may be removed as soon as 9 months after their deprecation was announced,
provided that our documentation and dogfooding do not use them.