Add jerop and priya to governing board, remove andrew 🗝️
1. Finishing up updates from governing board election (andrew bayer
   no longer on the board, priya wadhwa joining the board)
2. @jerop will be taking my place on the governing board while I am
   on leave (see [leave policy](https://github.com/tektoncd/community/blob/main/governance.md#governing-board-leave-policy))
   (thanks @jerop !!)

Went through the list at https://github.com/tektoncd/community/blob/main/governance.md#changes-to-governing-board
to make sure jerop and priya have all the governing board permissions
and to make the changes required with andrew leaving the board.

Also added Priti to the governing board + community meeting faciltator
rotation because she is listed in the rotation in the doc.

While running the permissions adding script I ran into a couple of
difficulties which I've tried to fix in:
* https://github.com/tektoncd/plumbing/pull/1062
* https://github.com/tektoncd/plumbing/pull/1061