TEP-0108: Mapping Workspaces
This proposal builds on this prior work to reduce verbosity in
mapping `Workspaces` and improve usability of *Tekton Pipelines*.

`Tasks` declare `Workspaces` they need, while `Pipelines` declare
`Workspaces` that are shared among its `PipelineTasks`. The mapping
of `Workspaces` from `Pipelines` to `PipelineTasks` is verbose.

We propose auto-mapping `Workspaces` from `Pipelines` to `PipelineTasks`
when the names of the `Workspaces` declared in the `Pipeline` and
`PipelineTask` are the same.