TEP-0056: Pipelines in Pipelines [Problem Statement]
Today, users can define and execute `Tasks` and `Custom Tasks`
in `Pipelines`. In this TEP, we propose allowing users to define
and execute `Pipelines` in `Pipelines`, alongside `Tasks` and
`Custom Tasks`.

We previously described the motivation for this TEP in:
- https://github.com/tektoncd/community/pull/374
- https://github.com/tektoncd/community/pull/498

In this change, we build on that problem statement to include the
experiences and learnings using the experimental `Pipelines` in
`Pipelines` `Custom Task` over the past several months.

In summary, the motivation for `Pipelines` in `Pipelines` include
composability, reusability, failure strategies, skipping strategies,
data locality & pod overhead and secure software supply chain.