Update design principles guidance on conformance
This commit clarifies our design principles on conformance to specify
that proposals should in general make Tekton more platform-agnostic,
that existing Kubernetes-isms in the Tekton API shouldn't be used as examples to add more,
and that expecting the user to understand how the API is implemented is something of an
anti-pattern. This commit also updates the TEP template to include questions on conformance.