TEP-0110: Decouple Catalog Organization and Resource Reference
As described in TEP-0003: Tekton Catalog Organization,
the Tekton Catalog is a collection of blessed Tekton resources
(`Tasks` and `Pipelines`) that can be used with any system that
supports the Tekton API. On the hand, the Tekton Hub to be
able to search for resources.

Today, versioning of resources in the Catalog is directory-based
and there is tight coupling between the organization and how users
fetch resources. Therefore, it is challenging to change the
organization of the Catalog without breaking users.

This TEP proposes to make the *Hub* and any other **official**
mean of consuming Tekton resources from the community
decoupled from where we author them. In a gist, this aims to make
the repository `tektoncd/catalog` an implementation detail.
One approach is to support long-term and non-changing URI to refer
to resources via `https://` endpoints, such as the Hub or OCI images.
Another approach is to support remote resolution, through Hub Resolver
and Git Resolver, that does not depend on the organization of the
Catalogs.