TEP-0090: Matrix - Update Non Goals
[TEP-0090: Matrix][tep-0090] proposed executing a `PipelineTask` in
parallel `TaskRuns` and `Runs` with substitutions from combinations
of `Parameters` in a `Matrix`.

In this change, we update the non-goals of TEP-0090:

1. In non-goal 2, we indicated that controlling concurrency is out
of scope. We actually ended up addressing controlling the concurrency
of `TaskRuns` or `Runs` created in a given `Matrix`. See [details][cc]
in the TEP. In this change, we remove this non-goal.

2. In non-goals 4 and 5, we suggested that `when` expressions can be
used to guard the execution of a particular combination of `Parameters`
from a `Matrix`. This is not possible because `when` expressions guard
the execution of the whole `PipelineTask` (and its `TaskRuns`/`Runs`).
In this change, we update those non-goals to remove this suggestion.

[tep-0090]: https://github.com/tektoncd/community/blob/main/teps/0090-matrix.md
[cc]: https://github.com/tektoncd/community/blob/main/teps/0090-matrix.md#concurrency-control