TEP-0090: Matrix - Failure Strategies
[TEP-0090: Matrix][tep-0090] proposed executing a `PipelineTask` in
parallel `TaskRuns` and `Runs` with substitutions from combinations
of `Parameters` in a `Matrix`.

In this change, we propose a fail-fast stategy for handling failures
in `TaskRuns` and `Runs` created from a `Matrix`. This makes it easier
to control the execution of the many `TaskRuns` or `Runs` created from
a `Matrix` - up to 256.

This approach also aligns with the default behavior of `Matrix` in
other Continuous Delivery systems e.g [GitHub Actions][ghm-failfast].

If needed, we can explore supporting other failure strategies later.

[tep-0090]: https://github.com/tektoncd/community/blob/main/teps/0090-matrix.md
[ghm-failfast]: https://docs.github.com/en/actions/using-jobs/using-a-matrix-for-your-jobs#handling-failures