TEP-0090: Matrix - Status
[TEP-0090: Matrix][tep-0090] proposed executing a `PipelineTask` in
parallel `TaskRuns` and `Runs` with substitutions from combinations
of `Parameters` in a `Matrix`.

In this change, we describe how the status of a `PipelineRun` with
a matrixed `PipelineTask` would be in full and minimal modes.

Previously, we had included `matrixId` to the `PipelineRun` status
so that it can help with identifying `Results` from fanned out
`TaskRuns` or `Runs`. However, we ended up descoping that problem
so that it can be addressed when TEP-0075 and TEP-0076 are implemented.
See [`Results` from fanned-out `PipelineTasks`][results] for details.
In this change, we remove the `matrixId` in `PipelineRun` status so
that the problem can be addressed completely later.

[tep-0090]: https://github.com/tektoncd/community/blob/main/teps/0090-matrix.md
[results]: https://github.com/tektoncd/community/blob/main/teps/0090-matrix.md#results-from-fanned-out-pipelinetasks