TEP-0086: Larger Results via Sidecar Logs
We propose that we provide the multiple solutions, all guarded
behind a `larger-results` feature flag, so that we can experiment
and figure out a way forward. These gated solutions will be alpha,
and can be changed or removed at any time.

In this change, we propose experimenting with Sidecar Logs as a
solution for providing larger Results within the CRDs. This will
be enabled by setting `larger-results`: `"sidecar-logs"`.

This solution can be changed at any time, or removed completely.
This will give us an opportunity to gather user feedback and find
ways to address the concerns, as we figure out a way forward.

/kind tep