[TEP-0096] Decouple API and software versioning
This commit updates TEP-0096 to make it more clear which components of the TEP
are related to achieving "stable" stability levels for individual CRDs, and which
components are related to having a "V1" project. The intent of this commit is to
allow Tekton to release v1 versions of our CRDs without being blocked on metrics,
performance, or documentation issues, and to allow us to release v1 versions of these
CRDs at different points in time.