[TEP-0086] Add details for OIDC Storage API
This adds high level details about the OIDC Storage API that's been
discussed in WGs (and in various places in this doc already). This is
not intended to be a comprehensive design. API definitions, what storage
backend to use for the default impl, whether to support remote result
references, etc are intentionally not covered - we should be able to
adapt this design to whatever decision we eventually make.

Signed-off-by: Billy Lynch <billy@chainguard.dev>