TEP-0118: Matrix with Explicit Combinations of Parameters
[TEP-0090: Matrix][0090] introduced `Matrix` to the
`PipelineTask` specification such that the `PipelineTask`
executes a list of `TaskRuns` or `Runs` in parallel with the
specified list of inputs for a `Parameter` or with different
combinations of the inputs for a set of `Parameters`.

In this TEP, we build on the work in [TEP-0090: Matrix][0090]
by allowing passing in a specific combinations
of `Parameters` into the `Matrix`.

[0090]: https://github.com/tektoncd/community/blob/ccf53cfd58e037e99fbd846d4e7fd7b78a646a36/teps/0090-matrix.md

Co-authored-by: pritidesai <pdesai@us.ibm.com>