[TEP-0104]: Update case with limits but no requests
This commit focuses on the behavior of Task-level resource requirements
in the case where a user specifies Task-level limits but not Task-level
requests. In this case, we must set some requests, as Kubernetes' default
behavior will result in overall requests that are too large. This commit
proposes setting Task-level requests equal to Task-level limits, similar
to the behavior used by Kubernetes.