TEP-0075: More flexible ways to provide values for object param
related to https://github.com/tektoncd/pipeline/pull/5427.

Prior to this update, merge behaviour was not allowed between `default` and
run level value for object keys.

In that proposal and the implementation pr, we allow the merge behaviour
to happen to give users more flexible ways to provide values for object
param as long as all keys are provided with a value at runtime (combination
of the values provided from default and run level value).

Signed-off-by: Chuang Wang <chuangw@google.com>