Clarify LTS, support and number of releases
The policy covers support for releases but it does not explicitly
mention LTS anywhere, term which is used in releases.md files
across Tekton projects. Introduce the LTS term and explain how it
relates to the policy.

The term support is not clarified today - it implicitly aligns to
what we do today, so make that explicit in the policy and list
reason that justify a new patch release.

The policy today talks about four supported releases which matches
the Jan/Apr/Jul/Oct used by several projects in their releases.md.
To match that we shall do an LTS release every three months, not
four. Kubernets does three minor/major releases per year, so in
fact this policy does not align to the kubernets one, thus remove
the sentece about that.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>