TEP-0114: Custom Tasks Beta - Retries
In [TEP-0121][tep-0121], we are redesigining `Retries` in `TaskRuns`
and `CustomRuns`.

In this change, we propose excluding `retries` and `retriesStatus`
in the initial release of `CustomRuns`.

We will figure out a way forward in [TEP-0121][tep-0121], which could
be reintroducing the fields implementing replacement fields.

This allows us to unblock release of Custom Tasks Beta and Tekton
Pipelines V1.

/kind tep

[tep-0121]: https://github.com/tektoncd/community/pull/816