[TEP-0098] Update Problem Statement
This commit updates the Workflows problem statement to reflect updated goals of the project,
adding more detailed use cases and references to related projects or designs.
In particular, Remote Resolution has been implemented since this TEP was created,
so referencing Pipelines and Tasks from version control is less relevant to the TEP now.
It also adds detail on extensibility and conformance goals.