[TEP-0115] Use semVer in Catalog
Prior, the Tekton Catalog uses the simple SemVer model (<major-version>.<minor-version>) to version the resources for simplicity in the directory-based versioning.

This commit updates the proposal to use the full SemVer model (<major-version>.<minor-version>.<patch-version>) as we switch to the git-based versioning, where bug fixes or security patches can go to the patch version instead of a new minor version. This update also resolves the inconsistency and confusion between the Artifact Hub (which uses the full SemVer versioning model) and Tekton Catalog versioning.

Signed-off-by: Quan Zhang <zhangquan@google.com>