[TEP-0079] Verified Catalogs & Trusted Resource Integration
This commit updates the "Verified Remote Resources" section of tep-0079, which proposes to show a "signed" badge in the Artifact Hub UI for the resources containing signatures that can be verified as designed in tep-0091. This commit also proposes the design of the release and branch setup for the Verified Catalog repositories to make sure every release only contains signed resources.