Update s3c wg facilitators and details 🎊
Happy new year! In most recent WG we discussed updating the facilitator
rotation b/c it felt out of date with who was attending and we also
started using the hackmd doc to define the list (which is also much
easier to update). Anyone interested in joining as a facilitator plz add
yourself to the list there.

Also updated some other details that were out of date (links to closed
issues and discussions)