Add tektoncd-catalog org to org.yaml
Part of [#889][#889]. This commit add the `tektoncd-catalog` to the `org.yaml` file so that the org, org member, team and team members can be automatically managed by Peribolos (which lets `tektoncd-catalog` org reaches the same level of automation as `tektoncd`).

The role assignment can be only done for the members that have already joined the new org. The role is assigned to each individual based on his/her current role in `tektoncd/catalog` repo.

This commit only adds teams for the [golang][golang] repo as the migration of the repo is under going. More teams will be added in later PRs once we start the migration for other Verified Catalogs.

[golang]: https://github.com/tektoncd-catalog/golang
[#889]: #889