Add process for adopting projects
This commit makes several changes:

- Makes it easier to create experimental projects.
Instead of 2 gov board members approving an initial issue and then later a PR,
a contributor can just create a PR directly,
and merging that PR is sufficient for approval.

- Attempting to define conformance requirements, something that seems
important to the gov board but that we haven't defined yet

- Clarifying how an initial project proposal can become either a top level
project or an experimental project

- Having contributors write a TEP to create a top-level project, whether
it started as an experimental project, is adopted from another org,
or is adopted directly as a top level project. This reflects the fact
that almost all new Tekton projects since the TEP process started have
included a TEP, which was helpful in clarifying the project's goals.

- Adding a process for adopting projects. We currently state that a valid
path is for projects to start off in another repo and become a top level
repo later, but don't explain how this would happen. This commit
attempts to list the information we'd want about a project before deciding
whether to adopt it, and proposes putting that information in a TEP.

Co-authored-by: Jerop Kipruto <jerop@google.com>