Include opt-in step in election instructions
According to https://civs1.civs.us/faq.html#setting_up in Apr 2021
CIVS changed their system such that you have to opt into receiving
communication from the system in order to get the notification to
vote and/or to see the poll after opting in. Opting in can be done
at https://civs1.civs.us/cgi-bin/opt_in.pl.

Since this change was mad in Apr 2021 I feel a bit concerned about
whether we were actually able to reach all of the voters in the previous
election but 😅 nothing to be done about that now! (At least the outcome
was already known - which is in fact the case again and begs the
question of whether we should be doing the election at all BUT that's
another story :D)