Move inactive Pipelines reviewers to alumni (part 2)
The contributor ladder states that contributors may be moved to emeritus status due to extended periods of inactivity.
This commit moves Pipelines reviewers with no reviews in 2022 to alumni,
and alphabetizes the list of current reviewers.

To those who have been moved to alumni status, thank you for your contributions to this project,
and you are welcome to return to active reviewer status at any time.