Add `deferred` state for TEPs
Today, TEPs can have the following states: `proposed`,
`implementable`, `implemented`,`withdrawn` or `replaced`.

In this change, we add the `deferred` state for when a TEP
is not actively worked on. This is the only remaining state
we need to add from [KEP] process that inspired TEP process.

This state is useful when we want to move from `implementable`,
but `withdrawn` and `replaced` is not quite right because there
are some open questions. We ran into this scenario in [TEP-0092].

[KEP]: https://github.com/kubernetes/enhancements/blob/master/keps/sig-architecture/0000-kep-process/README.md#kep-workflow
[TEP-0092]: https://github.com/tektoncd/community/pull/969