Include authorship of pull requests in ladder requirements
Prior to this change, the *reviewer* and *maintainer* requirements
considered reviews only as the requirement. In this change, we include
authoring of pull requests as a meaningful contribution that counts
towards becoming a *reviewer* or *maintainer*.

Note that the *organization member* requirements already accounts for
both authoring and reviewing pull requests. In addition, the Kubernetes
[ladder] that inspired the Tekton ladder accounts for both authoring and
reviewing pull requests.

[ladder]: https://github.com/kubernetes/community/blob/master/community-membership.md#reviewer