Add design principles on CustomRun API
CustomRun is different from other Tekton APIs because its API
is implemented by custom run controllers, rather than in-tree
controllers. Therefore, adding new fields to this API should
be done with extra caution. This commit adds a design principle
providing guidance on whether a new field should go in the CustomRun
API or left to custom run controllers to support as part of the
custom spec or custom status, based on whether a feature is
applicable to all CustomRuns or just a feature of some of them.