TEP-0135: update coscheduling default value
Before this commit, the default value of `coscheduling` is `disabled`. This configuration does not preserve the current affinity assistant
behavior (coschedule workspaces). It also introduces confusion when configuring the new `coscheduling` and existing `disable-affinity-assistant`
feature flags together.

This commit update the default value of the `coscheduling` feature flag to `coschedule-workspaces`, which preserves the current affinity assistant
behavior. This commit also adds a chart of combinations of the `coscheduling` and the `disable-affinity-assistant` feature flags, and the corresponding
affinity assistant behaviors before and after migration.

/kind tep