TEP-0140: Produce Results from Matrixed PipelineTasks
[TEP-0090: Matrix][tep-0090] proposed executing a `PipelineTask`
in  parallel `TaskRuns` and `Runs` with substitutions from
combinations of `Parameters` in a `Matrix`. Specifying `Results`
in a `Matrix` was in scope in [TEP-0090][results], and is already
supported. However, producing `Results` from `PipelineTasks`
with a `Matrix` was out of scope to await
[TEP-0075: Object Parameters and Results][tep-0075] and
[TEP-0076: Array Results][tep-0076]. This TEP aims to enable
producing `Results` from `PipelineTasks` with a `Matrix` so
that they can be used in subsequent `PipelineTasks`.

[tep-0075]: https://github.com/tektoncd/community/blob/main/teps/0075-object-param-and-result-types.md
[tep-0076]: https://github.com/tektoncd/community/blob/main/teps/0076-array-result-types.md
[tep-0090]: https://github.com/tektoncd/community/blob/main/teps/0090-matrix.md
[tep-0118]: https://github.com/tektoncd/community/blob/main/teps/0118-matrix-with-explicit-combinations-of-parameters.md

Signed-off-by: Priti Desai <pdesai@us.ibm.com>