TEP-0148: Summary/Aggregation API for Tekton Results
This TEP proposes a new API endpoint for Tekton results that will allow users to
get a summary or aggregation data given a set of filters. This will be useful
for getting a quick overview without having to go through all the response and also
reduce the amount of data that needs to be transferred. This will also provide
a way to get aggregated data for a set of filters.

Co-authored-by: Khurram Baig <kbaig@redhat.com>
Signed-off-by: Avinal Kumar <avinal@redhat.com>