[TEP-0144] Param Subset Enforcement
Prior to this commit, Tekton validates the intersection of the enums that are both specified at `Pipeline` level and at the referenced `PipelineTask` level.
In this solution, users are burdened with calculating intersections by iterating through all the PipelineTasks.

This commits add an extra validation to require that the pipeline-level enum must be a subset of the corresponding pipeline-task level enum. With this approach,
users only need to check the pipeline-level enum to execute the pipeline successfully.

/kind tep